requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==3.2.23
djangorestframework==3.15.1
geographiclib==2.0
geopy==2.4.1
idna==3.10
numpy==2.3.0
pandas==2.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
